HackerNewsUI/src/app/app.component.ts
import { CUSTOM_ELEMENTS_SCHEMA, Component } from '@angular/core';
import { CommonModule } from '@angular/common';
import { RouterOutlet } from '@angular/router';
import { HackerNewsStoryComponent } from './hacker-news-story/hacker-news-story.component';
import { HttpService } from './_services/http.services';
import { HttpClientModule } from '@angular/common/http';

@Component({
  selector: 'app-root',
  standalone: true,
  imports: [CommonModule, RouterOutlet,HttpClientModule,HackerNewsStoryComponent],
  providers:[HttpService],
  templateUrl: './app.component.html',
  styleUrl: './app.component.scss',
  schemas:[CUSTOM_ELEMENTS_SCHEMA]
})
export class AppComponent {
  title = 'HackerNews';
}
